Landlord Service Controller

+++ src/main/java/com/example/myApp/apartment/dtos/ApartmentResponseDto.java
@@ -19,7 +19,7 @@ public class ApartmentResponseDto {
   private String postalCode;
   private String apartmentNumber;
   private Integer floor;
-  private BigDecimal areaSqm;
+  private Double areaSqm;
   private String buildingRegNumber;
   private String cadastralArea;
   private String titleDeedNumber;
